prod/backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.4
psycopg[binary]>=3.2
requests>=2.31,<3